backend/requirements.txt
asgiref==3.9.2
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
google-auth==2.40.3
google-auth-oauthlib==1.2.2
idna==3.10
oauthlib==3.3.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.10.1
python-dotenv==1.1.1
razorpay==2.0.0
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
sqlparse==0.5.3
urllib3==2.5.0
